refactor: pages/ 폴더 제거 및 routes/ 하위로 이동
#6
