+++ src/routes/user/-components/Steps.module.scss
@@ -1,4 +1,4 @@
-@use 'src/styles/variables';
+@use '../../../styles/variables';
 
 .stepContainer {
   height: 100%;
