+++ src/routes/user/index.lazy.tsx
@@ -4,9 +4,9 @@ import { SweaterType } from "../../types/enum";
 import { useUpdateUser } from "../../hooks/useUsers.ts";
 import { UpdateUserRequest } from "../../types/api";
 import useUserStore from "../../stores/userStore.ts";
-import { Step1 } from "../../pages/Users/Step1.tsx";
-import { Step2 } from "../../pages/Users/Step2.tsx";
-import { Step3 } from "../../pages/Users/Step3.tsx";
+import { Step1 } from "./-components/Step1.tsx";
+import { Step2 } from "./-components/Step2.tsx";
+import { Step3 } from "./-components/Step3.tsx";
 import styles from './index.lazy.module.scss';
 import { CommonLayout } from "../../layouts/CommonLayout.tsx";
 import { useToastStore } from "../../stores/toastStore.ts";
